Separate library files and test files.
