+++ lib/array.php
