+++ lib/general.php
