+++ lib/string.php
